+++ src/test/java/step2/CalculatorTest.java
@@ -15,34 +15,6 @@ class CalculatorTest {
         calculator = new Calculator();
     }
 
-    @Test
-    @DisplayName("덧셈")
-    public void add() {
-        Integer value =  OperatorType.searchOperation("+").calculate(1, 3);
-        assertEquals(value, 4);
-    }
-
-    @Test
-    @DisplayName("뺄셈")
-    public void minus() {
-        Integer value =  OperatorType.searchOperation("-").calculate(1, 2);
-        assertEquals(value, -1);
-    }
-
-    @Test
-    @DisplayName("곱셈")
-    public void multipication() {
-        Integer value =  OperatorType.searchOperation("*").calculate(2, 2);
-        assertEquals(value, 4);
-    }
-
-    @Test
-    @DisplayName("나눗셈")
-    public void divide() {
-        Integer value =  OperatorType.searchOperation("/").calculate(5, 2);
-        assertEquals(value, 2);
-    }
-
     @DisplayName("입력 값이 빈공백, null일 경우 IllegalArgumentException throw")
     @ParameterizedTest
     @NullAndEmptySource
@@ -63,11 +35,8 @@ class CalculatorTest {
     public void implementationFunction(String input, Integer result) {
         assertEquals(calculator.calculate(input), result);
     }
-
-
-
-
-
-
-
 }
+
+
+
+
